feat: Update ComplementAddBaseForm to include slug parameter
The code changes modify the ComplementAddBaseForm component to include a slug parameter. This parameter is passed as a prop to the component and is used to set the slug value in the form. This enhancement improves the functionality of the codebase by allowing the form to dynamically handle different slugs when adding complements.