refactor: Fix error handling in createClass mutation
The createClass mutation in classes.ts has been updated to fix the error handling. Previously, the return value for the "id" field was a string, but it has been changed to a boolean. This change ensures that the error flag is correctly set when there is an error creating the class.