refactor: Update ClasseServerLayer to fix notes link
The ClasseServerLayer component in ClasseServerLayer.tsx has been updated to fix the notes link. The href has been changed from `/sequences/notes/${props.slug}` to `/classes/notes/${props.slug}` to ensure that the correct page is linked. This change improves the accuracy of the link and provides a better user experience.