refactor: Add sequence class mutation and queries
This commit adds a new mutation and queries related to sequence classes. The `addSequenceClass` mutation allows adding a sequence to a class by specifying the class ID and sequence ID. The `getSequencesClass` query retrieves all sequences associated with a class based on the class ID. The `deleteSequenceClass` mutation deletes a sequence from a class by providing the sequence ID. These changes enhance the functionality of the application by allowing better management of sequences in classes.