feat: Add useCreateBaseEvaluation hook for creating base evaluations
This commit adds a new hook called useCreateBaseEvaluation. The hook provides a mutation function for creating base evaluations. It utilizes the useMutation hook from the @tanstack/react-query library to handle the mutation logic. The mutation function calls the createEvaluation method from the evaluationUsecases module to create the evaluation. If the evaluation creation is successful, the hook returns the mutation object. If there is an error, the hook displays a toast error message and throws an error. This addition improves code organization and separation of concerns.