feat: Add criterias and isGraded fields to ConvexDatabase constructor
This commit adds the criterias and isGraded fields to the ConvexDatabase constructor in the convex-impl.ts file. The criterias field is used to store the evaluation criteria, while the isGraded field indicates whether the evaluation is graded or not. This addition enhances the functionality of the ConvexDatabase class by allowing for more detailed evaluation configurations.