feat: Add getEvaluation function
This commit adds a new function called getEvaluation to the evaluation application adapters. The function is responsible for fetching an evaluation based on the provided evaluationId. It uses the evaluationUsecases.getEvaluationBase function to retrieve the evaluation data. This addition enhances the functionality of the application by providing a way to fetch evaluation data for further processing.