feat: Add lodash dependency to useCreateEvaluationBaseFormLogic.ts
This commit adds the lodash dependency to the useCreateEvaluationBaseFormLogic.ts file. The lodash library is used to import the `add` function, which is used in the logic of the `useCreateEvaluationBaseFormLogic` hook. This addition enhances the functionality of the hook by allowing the use of the `add` function for array manipulation.