feat: Add @tiptap/extension-link to package.json and pnpm-lock.yaml
This commit adds the "@tiptap/extension-link" package to the package.json file and updates the pnpm-lock.yaml file to include the necessary dependencies and versions. The "@tiptap/extension-link" package is required for the functionality of the application and enhances the capabilities of the text editor by allowing the insertion of hyperlinks. This addition improves the overall user experience by providing a more comprehensive set of editing features.