refactor: Add checkSpecialGradeType utility function
A new utility function, checkSpecialGradeType, has been added to the project. This function checks if a given grade is one of the special grade types defined in the SpecialGradeTypes union. It returns an object with a boolean shouldReturn property indicating whether the grade is a special type or not, and a returnValue property containing the grade value. This addition enhances the flexibility and accuracy of grade type checking in the evaluation application.