refactor: Update TenPointsCriteriaForm to fix grade onChange bug
This commit fixes a bug in the TenPointsCriteriaForm component where the grade onChange event was not correctly updating the field value. The bug was causing the grade to always be set to the maximum value instead of the user input. This commit updates the onChange event handler to correctly set the field value based on the user input, ensuring that the grade is accurately reflected in the form.