refactor: Update useGetEvaluationCompoundList to enable/disable query based on options.enabled
This commit updates the useGetEvaluationCompoundList hook to include an options.enabled parameter. The enabled parameter allows the developer to control whether the query should be executed or not. By default, the query is enabled, but it can be disabled by setting enabled to false. This enhancement provides more flexibility and control over when the evaluation compound list should be fetched.