refactor: Update UpdateStudentGradeForm to fetch compound list based on options.enabled
This commit updates the UpdateStudentGradeForm component to fetch the compound list based on the value of the options.enabled parameter. The useGetEvaluationCompoundList hook now accepts the enabled parameter, allowing the developer to control whether the query should be executed or not. This enhancement provides more flexibility and control over when the evaluation compound list should be fetched.