refactor: Update LayoutWithProps component to conditionally render ErrorDialog component
This commit updates the LayoutWithProps component in the LayoutWithProps.tsx file to include conditional rendering of the ErrorDialog component. If the isError prop is truthy, the ErrorDialog component will be rendered with the provided props. This change enhances the flexibility of the component and improves the handling of error scenarios.