refactor: Update EditorProvider and HeadingMenuBar to support setting heading level with optional position
This commit updates the EditorProvider and HeadingMenuBar components to support setting the heading level with an optional position parameter. In the EditorProvider component, the useEditor hook from "@tiptap/react" is imported to enable the usage of the editor. In the HeadingMenuBar component, the setHeading function is modified to accept an additional position parameter, which allows for setting the heading level at a specific position in the editor. This enhancement improves the flexibility and usability of the editor by providing more control over the heading levels and their positions.