refactor: Update FloatingMenuBar to use MenuTextColor component with editor prop
This commit updates the FloatingMenuBar component to use the MenuTextColor component with the editor prop. Previously, the MenuTextColor component was used without the editor prop, which caused a functional issue. By passing the editor prop, we ensure that the MenuTextColor component has access to the necessary data and functions from the editor. This change improves the functionality and reliability of the FloatingMenuBar component.