refactor: Update MenuImage component to receive editor prop
This commit modifies the MenuImage component to receive the editor prop. Previously, the component relied on the useCurrentEditor hook to access the editor instance, which caused potential issues when the hook returned null. By passing the editor prop directly, we ensure that the MenuImage component always has access to the necessary editor data and functions. This change improves the reliability and maintainability of the MenuImage component.